build system improvements, clang-tidy integration for warnings equals errors (need to continue). POSIX socket legacy code removal

+++ include/yams/chunking/chunker.h
@@ -164,7 +164,7 @@ struct DeduplicationStats {
     size_t uniqueChunks = 0;
 
     double getRatio() const {
-        return totalSize > 0 ? 1.0 - (static_cast<double>(uniqueSize) / totalSize) : 0.0;
+        return totalSize > 0 ? 1.0 - (static_cast<double>(uniqueSize) / static_cast<double>(totalSize)) : 0.0;
     }
 };
